Fixed api Rest for get articles and added api for shared articles

+++ database/migrations/2013_06_11_100000_create_cities_table.php
@@ -15,7 +15,7 @@ class CreateCitiesTable extends Migration
     {
         Schema::create('cities', function (Blueprint $table) {
             $table->increments('id');
-            $table->smallInteger('state');
+            $table->Integer('state');
             $table->char('name', 32);
             //$table->foreign('state')->references('id')->on('states');
             $table->timestamps();
